Lint & Project File Reorganization
