+++ Mileage Tracker/Extensions/UIApplicationVersion.swift
@@ -1,6 +1,6 @@
 //
 //  UIApplicationVersion.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 23/03/2021.
 //
