+++ Mileage Tracker/Extensions/UIDeviceExtension.swift
@@ -1,6 +1,6 @@
 //
 //  UIDeviceExtension.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 23/03/2021.
 //
