+++ Mileage Tracker/Extensions/View+dateToString.swift
@@ -1,6 +1,6 @@
 //
 //  View+dateToString.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 25/03/2021.
 //
