+++ Mileage Tracker/Localization/Strings.swift
@@ -1,6 +1,6 @@
 //
 //  Strings.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by tarrask on 08/02/2023.
 //
