+++ Mileage Tracker/Model/OtherCost+FakeData.swift
@@ -1,6 +1,6 @@
 //
 //  OtherCost+FakeData.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by tarrask on 08/02/2023.
 //
