+++ Mileage Tracker/Model/OtherCosts.swift
@@ -1,6 +1,6 @@
 //
 //  OtherCosts.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 15/01/2021.
 //
