+++ Mileage Tracker/View/GraphicsView/Subviews/Graphics+GraphicSelector.swift
@@ -1,0 +1,22 @@
+//
+//  Graphics+Selector.swift
+//  Fuel Logue
+//
+//  Created by Alex Luna on 06/03/2024.
+//
+
+import SwiftUI
+
+extension GraphicsView {
+    @ViewBuilder
+    func graphicSelector() -> some View {
+        // MARK: - Graphics Selector
+        Picker(graphicString, selection: $graphType) {
+            Text(refuelString).tag(GraphType.spending)
+            Text("Price/"+settingsMG.chosenVolume).tag(GraphType.dates)
+            Text(settingsMG.chosenDistance+"/"+settingsMG.chosenCurrency).tag(GraphType.km)
+        }
+        .pickerStyle(SegmentedPickerStyle())
+        .padding(.horizontal, 20)
+    }
+}
