+++ Mileage Tracker/View/Helper Views/ShareSheet.swift
@@ -1,6 +1,6 @@
 //
 //  ShareSheet.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 24/03/2021.
 //
