+++ Mileage Tracker/View/Main Tab Views/MainAppView.swift
@@ -1,6 +1,6 @@
 //
 //  MainAppTabs.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 26/03/2021.
 //
@@ -10,22 +10,21 @@ import SwiftUI
 struct MainAppView: View {
     @EnvironmentObject var settingsMG: SettingsManager
 
-    @StateObject var tracker = MileageTracker.shared
+    @StateObject var mileageVM = MileageViewModel.shared
 
-    @StateObject var viewRouter = ViewRouter.shared
+    @StateObject var viewRouter = ViewRouter()
 
     var body: some View {
         VStack(spacing: 0) {
-            // MARK: - Main Page Views Switcher
             mainSwitcher()
-            // MARK: - Tab Bar
+
             tabBar()
         }
-        .environmentObject(tracker)
+        .environmentObject(mileageVM)
         .environmentObject(viewRouter)
         .edgesIgnoringSafeArea([.bottom, .horizontal])
     }
-    
+
     @ViewBuilder
     func mainSwitcher() -> some View {
         switch viewRouter.currentPage {
@@ -39,40 +38,6 @@ struct MainAppView: View {
             GraphicsView()
         case .settings:
             SettingsView()
-        }
-    }
-    
-    @ViewBuilder
-    func tabBar() -> some View {
-        ZStack {
-            HStack(spacing: 25) {
-                // MileageView
-                TabBarIcon(assignedPage: .mileage,
-                           systemIconName: "drop.fill",
-                           tabName: refuelsString)
-                
-                // OtherCostsView
-                TabBarIcon(assignedPage: .othercosts,
-                           systemIconName: "wrench.and.screwdriver",
-                           tabName: otherCostsString)
-                
-                // New Add Button
-                TabBarIcon(assignedPage: .addobject,
-                           systemIconName: "plus.circle.fill",
-                           tabName: addCostString)
-                
-                // StatsView
-                TabBarIcon(assignedPage: .graphics,
-                           systemIconName: "list.star",
-                           tabName: statsString)
-                
-                // SettingsView
-                TabBarIcon(assignedPage: .settings,
-                           systemIconName: "gearshape",
-                           tabName: settingsString)
-            }
-            .frame(maxWidth: .infinity, maxHeight: 70)
-            .background(settingsMG.theme.backgroundColor.shadow(radius: 2))
         }
     }
 }
