+++ Mileage Tracker/View/Onboarding/NewFeatureOnboarding.swift
@@ -1,6 +1,6 @@
 //
 //  NewFeatureOnboarding.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 29/03/2021.
 //
