+++ Mileage Tracker/View/Onboarding/OnboardingView.swift
@@ -1,6 +1,6 @@
 //
 //  OnboardingView.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 25/03/2021.
 //
