+++ Mileage Tracker/View/Onboarding/SalesOnboarding.swift
@@ -1,6 +1,6 @@
 //
 //  SalesOnboarding.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 29/03/2021.
 //
