+++ Mileage Tracker/View/Onboarding/WelcomeOnboarding.swift
@@ -1,6 +1,6 @@
 //
 //  WelcomeOnboarding.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 29/03/2021.
 //
