+++ Mileage Tracker/ViewModel/Settings Manager.swift
@@ -1,6 +1,6 @@
 //
 //  Theme Manager.swift
-//  Mileage Tracker
+// Fuel Logue
 //
 //  Created by Alex Luna on 25/03/2021.
 //
